Move algo to py_kwiver
