+++ py_kwiver/vital/algo/CMakeLists.txt
