+++ py_kwiver/vital/algo/algorithm.h
