+++ py_kwiver/vital/algo/algorithm_module.cxx
