+++ py_kwiver/vital/algo/image_object_detector.h
